+++ NEWS.md
@@ -1,3 +1,9 @@
+# tactile 0.1.0.9000
+## New panel functions
+* `panel.ci` plots confidence levels for an `xyplot`-type of plot, provided
+that confidence limits are given to arguments `lower` and `upper` in the
+high-level call.
+
 # tactile 0.1.0
 
 * This is the first CRAN release.